requirements.txt
tensorflow-macos==2.13.0  # Machine learning framework (macOS version)
Flask==3.0.0             # Web framework
Pillow==10.1.0          # Image processing library
numpy==1.24.3           # Numerical computing library